frontend

+++ src/components/Admin/AllBrands.jsx
@@ -5,8 +5,8 @@ import { Button } from "@material-ui/core";
 import axios from "axios";
 import { backend_url, server } from '../../server';
 import { AiOutlineDelete } from 'react-icons/ai';
-import Loader from "../Layout/Loader";
 import { toast } from "react-toastify";
+import { BsPencil } from 'react-icons/bs';
 
 const AllBrands = () => {
     const navigate = useNavigate();
@@ -17,23 +17,30 @@ const AllBrands = () => {
         axios.get(`${server}/brands/get-all-brands`, { withCredentials: true }).then((res) => {
             setBrands(res.data.brands);
             setIsLoading(false);
-        })
+        }).catch((e) => setIsLoading(false));
     }, []);
 
     const handleDelete = async (id) => {
-	await axios.delete(
-            `${server}/brands/delete-brand/${id}`,
-            { withCredentials: true, }
-        ).then(({data}) => {
-	   if (data.success) {
-             toast.success("Brand deleted successfully!");
-             navigate("/admin-brands");
-	     window.location.reload();
-           }
-	}).catch((error) => {
-	    toast.error(error.response.data.message);
-	});
+        let e = window.confirm("Do you want to Delete it?");
+        if (e) {
+            await axios.delete(
+                `${server}/brands/delete-brand/${id}`,
+                { withCredentials: true, }
+            ).then(({ data }) => {
+                if (data.success) {
+                    toast.success("Brand deleted successfully!");
+                    navigate("/admin-brands");
+                    window.location.reload();
+                }
+            }).catch((error) => {
+                toast.error(error.response.data.message);
+            });
+        }
     };
+
+    const handleEdit = (data) => {
+        navigate(`/admin-brand/new?_id=${data.id}&name=${data.name}&image=${data.image}`);
+    }
 
     const columns = [
         {
@@ -50,6 +57,23 @@ const AllBrands = () => {
             flex: 0.6,
             align: 'center',
             renderCell: (params) => <img className='w-auto h-auto object-contain border-none' src={`${backend_url}${params.value}`} alt='img' />,
+        },
+        {
+            field: "Edit",
+            flex: 0.4,
+            minWidth: 100,
+            headerName: "",
+            type: "number",
+            sortable: false,
+            renderCell: (params) => {
+                return (
+                    <>
+                        <Button onClick={() => handleEdit(params.row)}>
+                            <BsPencil size={20} />
+                        </Button>
+                    </>
+                );
+            },
         },
         {
             field: "Delete",
@@ -81,30 +105,32 @@ const AllBrands = () => {
             });
         });
 
+    if (isLoading) {
+        return (
+            <div className="flex justify-center items-center w-full h-screen">
+                <div className="rounded-full border-t-4 border-b-4 border-red-600 h-20 w-20 animate-spin"></div>
+            </div>
+        )
+    }
+
     return (
-        <>
-            {isLoading ? (
-                <Loader />
-            ) : (
-                <div className="w-full mx-8 pt-1 mt-10 bg-white">
-                    <div className='m-5 text-center'>
-                        <Link to={'/admin-brand/new'}>
-                            <Button variant="outlined">
-                                Add New Brand
-                            </Button>
-                        </Link>
-                    </div >
-                    <DataGrid
-                        rows={row}
-                        columns={columns}
-                        pageSize={2}
-                        disableSelectionOnClick
-                        autoHeight
-                        rowHeight={200}
-                    />
-                </div>
-            )}
-        </>
+        <div className="w-full mx-8 pt-1 mt-10 bg-white">
+            <div className='m-5 text-center'>
+                <Link to={'/admin-brand/new'}>
+                    <Button variant="outlined">
+                        Add New Brand
+                    </Button>
+                </Link>
+            </div >
+            <DataGrid
+                rows={row}
+                columns={columns}
+                pageSize={2}
+                disableSelectionOnClick
+                autoHeight
+                rowHeight={200}
+            />
+        </div>
     );
 
 }
