+++ src/components/Shop/AllEvents.jsx
@@ -3,12 +3,13 @@ import { DataGrid } from "@material-ui/data-grid";
 import React, { useEffect } from "react";
 import { AiOutlineDelete, AiOutlineEye } from "react-icons/ai";
 import { useDispatch, useSelector } from "react-redux";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { deleteEvent, getAllEventsShop } from "../../redux/actions/event";
-import Loader from "../Layout/Loader";
 import RenderExpandableCell from "../Layout/RenderExpandableCell";
+import { BsPencil } from 'react-icons/bs';
 
 const AllEvents = () => {
+  const navigate = useNavigate();
   const { events, isLoading } = useSelector((state) => state.events);
   const { seller } = useSelector((state) => state.seller);
 
@@ -16,11 +17,18 @@ const AllEvents = () => {
 
   useEffect(() => {
     dispatch(getAllEventsShop(seller._id));
-  }, [dispatch]);
+  }, [dispatch, seller._id]);
 
   const handleDelete = (id) => {
-    dispatch(deleteEvent(id));
-    window.location.reload();
+    let e = window.confirm("Do you want to Delete it?");
+    if (e) {
+      dispatch(deleteEvent(id));
+      window.location.reload();
+    }
+  }
+
+  const handleEdit = (data) => {
+    navigate(`/dashboard-create-event?_id=${data.id}`);
   }
 
   const columns = [
@@ -70,8 +78,6 @@ const AllEvents = () => {
       type: "number",
       sortable: false,
       renderCell: (params) => {
-        const d = params.row.name;
-        const product_name = d.replace(/\s+/g, "-");
         return (
           <>
             <Link to={`/product/${params.id}?isEvent=true`}>
@@ -79,6 +85,23 @@ const AllEvents = () => {
                 <AiOutlineEye size={20} />
               </Button>
             </Link>
+          </>
+        );
+      },
+    },
+    {
+      field: "Edit",
+      flex: 0.4,
+      minWidth: 100,
+      headerName: "",
+      type: "number",
+      sortable: false,
+      renderCell: (params) => {
+        return (
+          <>
+            <Button onClick={() => handleEdit(params.row)}>
+              <BsPencil size={20} />
+            </Button>
           </>
         );
       },
@@ -94,7 +117,7 @@ const AllEvents = () => {
         return (
           <>
             <Button
-            onClick={() => handleDelete(params.id)}
+              onClick={() => handleDelete(params.id)}
             >
               <AiOutlineDelete size={20} />
             </Button>
@@ -107,7 +130,7 @@ const AllEvents = () => {
   const row = [];
 
   events &&
-  events.forEach((item) => {
+    events.forEach((item) => {
       row.push({
         id: item._id,
         name: item.name,
@@ -117,22 +140,24 @@ const AllEvents = () => {
       });
     });
 
+  if (isLoading) {
+    return (
+      <div className="flex justify-center items-center w-full h-screen">
+        <div className="rounded-full border-t-4 border-b-4 border-red-600 h-20 w-20 animate-spin"></div>
+      </div>
+    )
+  }
+
   return (
-    <>
-      {isLoading ? (
-        <Loader />
-      ) : (
-        <div className="w-full mx-8 pt-1 mt-10 bg-white">
-          <DataGrid
-            rows={row}
-            columns={columns}
-            pageSize={8}
-            disableSelectionOnClick
-            autoHeight
-          />
-        </div>
-      )}
-    </>
+    <div className="w-full mx-8 pt-1 mt-10 bg-white">
+      <DataGrid
+        rows={row}
+        columns={columns}
+        pageSize={8}
+        disableSelectionOnClick
+        autoHeight
+      />
+    </div>
   );
 };
