+++ src/components/Shop/AllProducts.jsx
@@ -3,13 +3,14 @@ import { DataGrid } from "@material-ui/data-grid";
 import React, { useEffect } from "react";
 import { AiOutlineDelete, AiOutlineEye } from "react-icons/ai";
 import { useDispatch, useSelector } from "react-redux";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { getAllProductsShop } from "../../redux/actions/product";
 import { deleteProduct } from "../../redux/actions/product";
-import Loader from "../Layout/Loader";
 import RenderExpandableCell from "../Layout/RenderExpandableCell";
+import { BsPencil } from 'react-icons/bs';
 
 const AllProducts = () => {
+  const navigate = useNavigate();
   const { products, isLoading } = useSelector((state) => state.products);
   const { seller } = useSelector((state) => state.seller);
 
@@ -17,18 +18,25 @@ const AllProducts = () => {
 
   useEffect(() => {
     dispatch(getAllProductsShop(seller._id));
-  }, [dispatch]);
+  }, [dispatch, seller._id]);
 
   const handleDelete = (id) => {
-    dispatch(deleteProduct(id));
-    window.location.reload();
+    let e = window.confirm("Do you want to Delete it?");
+    if (e) {
+      dispatch(deleteProduct(id));
+      window.location.reload();
+    }
   };
+
+  const handleEdit = (data) => {
+    navigate(`/dashboard-create-product?_id=${data.id}`);
+  }
 
   const columns = [
     {
       field: "id",
       headerName: "Product Id",
-      minWidth: 130, 
+      minWidth: 130,
       flex: 0.8,
       renderCell: params => <RenderExpandableCell {...params} />
     },
@@ -83,6 +91,23 @@ const AllProducts = () => {
       },
     },
     {
+      field: "Edit",
+      flex: 0.4,
+      minWidth: 100,
+      headerName: "",
+      type: "number",
+      sortable: false,
+      renderCell: (params) => {
+        return (
+          <>
+            <Button onClick={() => handleEdit(params.row)}>
+              <BsPencil size={20} />
+            </Button>
+          </>
+        );
+      },
+    },
+    {
       field: "Delete",
       flex: 0.8,
       minWidth: 120,
@@ -114,22 +139,24 @@ const AllProducts = () => {
       });
     });
 
+  if (isLoading) {
+    return (
+      <div className="flex justify-center items-center w-full h-screen">
+        <div className="rounded-full border-t-4 border-b-4 border-red-600 h-20 w-20 animate-spin"></div>
+      </div>
+    )
+  }
+
   return (
-    <>
-      {isLoading ? (
-        <Loader />
-      ) : (
-        <div className="w-full mx-8 pt-1 mt-10 bg-white">
-          <DataGrid
-            rows={row}
-            columns={columns}
-            pageSize={8}
-            disableSelectionOnClick
-            autoHeight
-          />
-        </div>
-      )}
-    </>
+    <div className="w-full mx-8 pt-1 mt-10 bg-white">
+      <DataGrid
+        rows={row}
+        columns={columns}
+        pageSize={8}
+        disableSelectionOnClick
+        autoHeight
+      />
+    </div>
   );
 };
